Cypher/Signer traits for plugable crypto
