+++ .gitignore
@@ -2,3 +2,4 @@
 Cargo.lock
 *.swp
 .vscode/*
+.DS_Store
