+++ Cargo.toml
@@ -13,9 +13,6 @@ license = "Apache-2.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 base64-url = "1.4.8"
-# Biscuit dependancies
-biscuit = { version = "0.5", optional = true }
-ring = { version = "0.16", optional = true }
 # Raw crypto dependancies
 chacha20poly1305 = { version = "0.7", optional = true }
 aes-gcm = { version = "0.8", optional = true }
@@ -39,4 +36,3 @@ x25519-dalek = "1.1"
 [features]
 default = ["raw-crypto"]
 raw-crypto = ["chacha20poly1305", "aes-gcm", "k256", "p256", "ed25519-dalek"]
-jose-biscuit = ["biscuit", "ring"]
