+++ src/messages/headers.rs
@@ -128,8 +128,11 @@ impl JwmHeader {
     pub fn get_iv(&self) -> &[u8] {
         &self.iv
     }
-    // FIXME: complete the implementation
+    /// Setter of JOSE header properties to identify which signature alg used.
+    /// Modifies `typ` and `alg` headers.
+    ///
     pub fn as_signed(&mut self, alg: SignatureAlgorithm) {
+        self.typ = String::from("JWM");
         match alg {
             SignatureAlgorithm::EdDsa => {
                 self.alg = Some(String::from("EdDSA"));
@@ -142,10 +145,11 @@ impl JwmHeader {
             },
         }
     }
-    /// Setter of proper headers to identify which crypto alg and key type to use.
-    /// Modifies `enc` and `alg` headers.
+    /// Setter of JOSE header preperties to identify which crypto alg and key type used.
+    /// Modifies `enc`, `typ` and `alg` headers.
     ///
     pub fn as_encrypted(&mut self, alg: CryptoAlgorithm) {
+        self.typ = String::from("JWM");
         match alg {
             CryptoAlgorithm::A256GCM => { 
                 self.enc = Some("A256GCM".into());
