+++ src/messages/mod.rs
@@ -5,9 +5,6 @@ mod types;
 mod jwe;
 mod jws;
 
-#[cfg(feature = "jose-biscuit")]
-mod biscuit;
-
 #[cfg(feature = "raw-crypto")]
 mod raw;
 
@@ -17,9 +14,6 @@ pub use headers::*;
 pub use types::*;
 pub use jwe::*;
 pub use jws::*;
-
-#[cfg(feature = "jose-biscuit")]
-pub use biscuit::*;
 
 #[cfg(feature = "raw-crypto")]
 pub use raw::*;
